Move record update handlers into utils folder
